CA-202608: Trigger refresh after system resume
If the guest is suspended and then resumed or migrated, trigger a
refresh to repopulate the xenstore entries (since the guest will have a
new domid).

Without doing this, XAPI thinks that the guest does not have PV drivers
installed and will refuse to do a vbd-plug operation until then next
refresh by the guest agent.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>